ITS/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
tensorflow==2.15.0
websockets==12.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
schedule==1.2.0
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.8.2

# 爬虫相关依赖
beautifulsoup4==4.12.2
feedparser==6.0.10
lxml==4.9.3
requests==2.31.0
selenium==4.15.2
fake-useragent==1.4.0
